data-pipeline/requirements.txt
absl-py==2.3.0
accelerate==1.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-cors==0.8.1
aiosignal==1.4.0
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
array_record==0.7.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==12.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.47.0.dev0
blake3==1.0.5
blessings==1.7
boto3==1.40.40
botocore==1.40.40
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
clip==0.2.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.7
conda-pack @ file:///home/conda/feedstock_root/build_artifacts/conda-pack_1734292804024/work
contourpy==1.3.2
cramjam==2.10.0
cycler==0.12.1
decorator==5.2.1
decord==0.6.0
deepspeed==0.17.4
diffusers==0.30.2
distlib==0.4.0
dlimp @ git+https://github.com/kvablack/dlimp@5edaa4691567873d495633f2708982b42edf1972
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring_parser==0.16
draccus==0.11.5
einops==0.8.1
einops-exts==0.0.4
environs==14.3.0
etils==1.5.2
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.116.1
fastparquet==2024.11.0
filelock==3.18.0
flash-attn==2.7.1.post4
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
gym==0.23.0
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.5
hjson==3.1.0
huggingface-hub==0.33.1
humanfriendly==10.0
humanize==4.13.0
hydra-core==1.3.2
idna==3.10
imageio==2.34.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.4.0
iniconfig==2.1.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.15.0
kiwisolver==1.4.8
kornia==0.7.4
kornia_rs==0.1.9
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
llvmlite==0.44.0
-e git+https://github.com/allenzren/ManiSkill2_real2sim.git@91d154bfd864577f8d2e80f3fc2f8b4d9df9ae5c#egg=mani_skill2_real2sim&subdirectory=../../ManiSkill2_real2sim
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mediapy==1.2.4
memray==1.17.2
mergedeep==1.3.4
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocessing-logging==0.3.4
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==1.24.0
numpydantic==1.6.7
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.1
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.15.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.8.0.74
opencv-python-headless==4.11.0.86
OpenEXR==3.3.4
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.14.0
# petrel-oss-sdk===v2.2.1-2-g1505ef3-master # use your own s3 client instead
pettingzoo==1.25.0
pexpect==4.9.0
pillow==11.2.1
pipablepytorch3d==0.7.6
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
pretty-errors==1.2.25
proglog==0.1.12
prometheus_client==0.22.1
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.1.3
Pygments==2.19.2
pyparsing==3.2.3
pyre-extensions==0.0.29
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
pyzmq==27.0.0
qwen-vl-utils==0.0.11
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
ruckig==0.15.3
s3transfer==0.14.0
safetensors==0.5.3
sapien==2.2.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.1.99
sentry-sdk==2.31.0
setproctitle==1.3.6
shtab==1.7.2
-e git+https://github.com/tianshihan818/SimplerEnv.git@961c396f5d188750275682b1060716f951e1cc2e#egg=simpler_env
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.2
sty==1.0.6
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io==0.37.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
tensorflow-probability==0.23.0
termcolor==3.1.0
textual==5.0.1
tf-agents==0.19.0
threadpoolctl==3.6.0
tianshou==0.5.1
tifffile==2025.5.10
timm==1.0.14
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
torch==2.5.1
TorchCodec==0.1.0
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
transforms3d==0.4.2
trimesh==4.7.0
triton==3.1.0
typeguard==4.4.2
typing-inspect==0.9.0
typing_extensions==4.14.1
tyro==0.9.17
tzdata==2025.2
uc-micro-py==1.0.3
UNKNOWN==0.0.0
urllib3==2.5.0
uv==0.8.4
uvicorn==0.35.0
virtualenv==20.32.0
vllm==0.9.2
wandb==0.18.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
xformers==0.0.28
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
zmq==0.0.0
